+++ tests/common/Widget/RadioListTest.php
@@ -111,7 +111,7 @@ final class RadioListTest extends TestCase
                     1 => 'One',
                     2 => 'Two',
                 ])
-                ->radioAttributes(['class' => 'red'])
+                ->replaceRadioAttributes(['class' => 'red'])
                 ->withoutContainer()
                 ->render(),
         );
@@ -128,7 +128,7 @@ final class RadioListTest extends TestCase
                     2 => 'Two',
                 ])
                 ->readonly()
-                ->radioAttributes(['class' => 'red'])
+                ->addRadioAttributes(['class' => 'red'])
                 ->withoutContainer()
                 ->render(),
         );
@@ -163,7 +163,7 @@ final class RadioListTest extends TestCase
                     2 => 'Two',
                     3 => 'Three',
                 ])
-                ->radioAttributes(['class' => 'red'])
+                ->replaceRadioAttributes(['class' => 'red'])
                 ->individualInputAttributes([
                     2 => ['class' => 'blue'],
                     3 => ['class' => 'green'],
@@ -185,7 +185,7 @@ final class RadioListTest extends TestCase
                     2 => 'Two',
                 ])
                 ->uncheckValue(0)
-                ->radioAttributes(['class' => 'red'])
+                ->replaceRadioAttributes(['class' => 'red'])
                 ->individualInputAttributes([
                     0 => ['class' => 'blue'],
                 ])
@@ -206,7 +206,7 @@ final class RadioListTest extends TestCase
                     2 => 'Two',
                     3 => 'Three',
                 ])
-                ->radioAttributes(['class' => 'red'])
+                ->replaceRadioAttributes(['class' => 'red'])
                 ->individualInputAttributes([
                     2 => ['class' => 'blue'],
                     3 => ['class' => 'green'],
@@ -232,7 +232,7 @@ final class RadioListTest extends TestCase
                     2 => 'Two',
                     3 => 'Three',
                 ])
-                ->radioAttributes(['class' => 'red'])
+                ->replaceRadioAttributes(['class' => 'red'])
                 ->individualInputAttributes([
                     2 => ['class' => 'blue'],
                     3 => ['class' => 'green'],
@@ -655,6 +655,7 @@ final class RadioListTest extends TestCase
         $this->assertNotSame($widget, $widget->containerTag(''));
         $this->assertNotSame($widget, $widget->containerAttributes([]));
         $this->assertNotSame($widget, $widget->radioAttributes([]));
+        $this->assertNotSame($widget, $widget->addRadioAttributes([]));
         $this->assertNotSame($widget, $widget->replaceRadioAttributes([]));
         $this->assertNotSame($widget, $widget->individualInputAttributes([]));
         $this->assertNotSame($widget, $widget->replaceIndividualInputAttributes([]));